revolutionary_time_crystal/requirements.txt
# Time-Crystal Photonic Isolator - Python Dependencies
# ===================================================
# 
# This file specifies exact package versions for reproducible builds.
# Generated: July 2025
# Target: Python 3.11+

# Core scientific computing
numpy>=1.24.0,<2.0.0
scipy>=1.10.0,<2.0.0
matplotlib>=3.6.0,<4.0.0
h5py>=3.8.0,<4.0.0

# Physics simulation (optional - graceful degradation if missing)
# Install with: conda install -c conda-forge pymeeus
# meep>=1.27.0

# Quantum computing (optional)
# qutip>=4.7.0

# Machine learning (optional)
# torch>=2.0.0
# torchvision>=0.15.0

# Parallel computing and optimization
psutil>=5.9.0
joblib>=1.3.0

# Data processing and analysis
pandas>=2.0.0,<3.0.0
seaborn>=0.12.0

# Development and testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-benchmark>=4.0.0
pytest-xdist>=3.3.0  # Parallel testing
black>=23.0.0  # Code formatting
flake8>=6.0.0  # Linting
mypy>=1.5.0  # Type checking

# Progress tracking and visualization
tqdm>=4.64.0
wandb>=0.12.0

# Physics simulation (mock for MEEP)
# meep>=1.24.0  # Uncomment for real MEEP installation
# python-meeus>=0.5.0

# Optimization and numerics
scikit-learn>=1.1.0
scikit-image>=0.19.0
cvxpy>=1.2.0
cvxopt>=1.3.0

# Parallel processing
multiprocessing-logging>=0.3.0
joblib>=1.1.0

# Quantum computing simulation
# qutip>=4.7.0  # Uncomment for full quantum simulation
# cirq>=1.0.0

# Development and testing
pytest>=7.0.0
pytest-cov>=3.0.0
black>=22.0.0
flake8>=4.0.0
mypy>=0.950

# Documentation
sphinx>=4.5.0
sphinx-rtd-theme>=1.0.0

# Jupyter for analysis
jupyter>=1.0.0
ipykernel>=6.0.0
ipywidgets>=7.6.0

# Additional scientific libraries
sympy>=1.10.0
astropy>=5.0.0
networkx>=2.8.0

# Performance optimization
numba>=0.56.0
cython>=0.29.0

# Image processing
pillow>=9.0.0
opencv-python>=4.5.0

# Statistical analysis
statsmodels>=0.13.0
pingouin>=0.5.0

# File format support
netcdf4>=1.5.0
xarray>=2022.3.0

# Molecular dynamics (for materials simulation)
ase>=3.22.0
pymatgen>=2022.4.0

# Signal processing
librosa>=0.9.0
pywavelets>=1.3.0

# Parallel computing
dask>=2022.4.0
distributed>=2022.4.0

# GPU acceleration support
cupy-cuda11x>=10.0.0  # For CUDA 11.x
# cupy-cuda12x>=12.0.0  # For CUDA 12.x (alternative)

# Advanced plotting
plotly>=5.7.0
bokeh>=2.4.0

# Memory profiling
memory-profiler>=0.60.0
psutil>=5.9.0

# Configuration management
pyyaml>=6.0
configparser>=5.2.0
hydra-core>=1.1.0

# Time series analysis
statsforecast>=1.0.0
prophet>=1.1.0

# Advanced optimization
optuna>=2.10.0
hyperopt>=0.2.0
ray[tune]>=1.13.0

# Scientific constants and units
pint>=0.19.0
astropy>=5.0.0

# Random number generation
randomgen>=1.21.0

# Advanced linear algebra
petsc>=3.17.0
petsc4py>=3.17.0

# Mesh generation and FEM
fenics>=2019.1.0
meshio>=5.3.0

# Cloud storage integration
boto3>=1.24.0
google-cloud-storage>=2.3.0

# API and web interfaces
fastapi>=0.75.0
uvicorn>=0.17.0
requests>=2.27.0

# Database connectivity
sqlalchemy>=1.4.0
pymongo>=4.1.0

# Compression
lz4>=4.0.0
zstandard>=0.17.0

# Advanced signal processing
spectrum>=0.8.0
pyfilterbank>=0.1.0

# Electromagnetic simulation helpers
python-EM>=0.1.0
pymaxwell>=0.1.0

# Mathematical optimization
gekko>=1.0.0
casadi>=3.5.0

# Advanced statistics
arviz>=0.12.0
pymc>=4.0.0

# Time handling
arrow>=1.2.0
pendulum>=2.1.0

# String processing
fuzzywuzzy>=0.18.0
python-levenshtein>=0.12.0

# System monitoring
psutil>=5.9.0
nvidia-ml-py3>=7.352.0

# Advanced file I/O
pyarrow>=8.0.0
fastparquet>=0.8.0

# Serialization
pickle5>=0.0.12
cloudpickle>=2.1.0

# Testing utilities
hypothesis>=6.46.0
factory-boy>=3.2.0

# Code quality
isort>=5.10.0
pre-commit>=2.19.0

# Documentation generation
mkdocs>=1.3.0
mkdocs-material>=8.2.0

# Performance benchmarking
py-spy>=0.3.0
line-profiler>=3.5.0

# Distributed computing
mpi4py>=3.1.0
horovod>=0.24.0

# Advanced plotting styles
tueplots>=0.0.4
scientific-visualization>=0.1.0
